+++ javacup-dashboard/src/main/java/io/cutalab/javacup/dashboard/views/CurrentJvmMetricsView.java
@@ -57,8 +57,7 @@ public class CurrentJvmMetricsView extends VerticalLayout {
         setSizeFull();
         setPadding(true);
         setSpacing(true);
-        getStyle().set("padding-bottom", "var(--lumo-space-xl)");
-        getStyle().set("padding-bottom", "var(--lumo-space-xl)");
+        getStyle().set("padding-bottom", "calc(var(--lumo-space-xl) * 2)");
 
         H1 title = new H1("Current JVM metrics");
         Paragraph description = new Paragraph(