+++ javacup-dashboard/src/main/java/io/cutalab/javacup/dashboard/views/MainView.java
@@ -16,8 +16,7 @@ public class MainView extends VerticalLayout {
         setSizeFull();
         setPadding(true);
         setSpacing(true);
-        getStyle().set("padding-bottom", "var(--lumo-space-xl)");
-        getStyle().set("padding-bottom", "var(--lumo-space-xl)");
+        getStyle().set("padding-bottom", "calc(var(--lumo-space-xl) * 2)");
 
         H1 title = new H1(AppInfo.NAME);
         Paragraph tagline = new Paragraph(AppInfo.TAGLINE);