+++ javacup-dashboard/src/main/java/io/cutalab/javacup/dashboard/views/ProcessesView.java
@@ -33,8 +33,7 @@ public class ProcessesView extends VerticalLayout {
         setSizeFull();
         setPadding(true);
         setSpacing(true);
-        getStyle().set("padding-bottom", "var(--lumo-space-xl)");
-        getStyle().set("padding-bottom", "var(--lumo-space-xl)");
+        getStyle().set("padding-bottom", "calc(var(--lumo-space-xl) * 2)");
 
         H1 title = new H1("Java processes");
         Paragraph description = new Paragraph(